rename source files and add index entry point
