+++ src/cli.test.ts
@@ -3,8 +3,8 @@ import { expect, test, vi } from "vitest";
 test("prints the fibonacci sequence to stdout", async () => {
   const spy = vi.spyOn(process.stdout, "write").mockReturnValue(true);
 
-  process.argv = ["node", "bin.js", "5"];
-  await import("./bin.js");
+  process.argv = ["node", "cli.js", "5"];
+  await import("./cli.js");
 
   expect(spy).toHaveBeenCalledExactlyOnceWith("1 1 2 3 5\n");
 });
